refactor(Direction): 변수명 변경
to -> compact
